Restructure CLI (#223)
* Initial changes

* Finish relocating files

* Add __init__

* Move args to separate lines

* Delete conda
