--- conda/meta.yaml
@@ -1,70 +1,0 @@
-{% set name = 'sdgym' %}
-{% set version = '0.6.1.dev0' %}
-
-package:
-  name: "{{ name|lower }}"
-  version: "{{ version }}"
-
-source:
-  url: "https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.tar.gz"
-
-build:
-  number: 0
-  noarch: python
-  entry_points:
-    - sdgym=sdgym.__main__:main
-  script: "{{ PYTHON }} -m pip install . -vv"
-
-requirements:
-  host:
-    - pip
-    - pytest-runner
-    - graphviz
-    - python >=3.6,<3.10
-    - appdirs >=1.3,<2
-    - boto3 >=1.15.0,<2
-    - botocore >=1.18,<2
-    - humanfriendly >=8.2,<11
-    - numpy >=1.18.0,<2
-    - pandas >=1.1.3,<2
-    - pomegranate >=0.13.4,<15
-    - psutil >=5.7,<6
-    - scikit-learn >=0.24,<2
-    - tabulate >=0.8.3,<0.9
-    - pytorch >=1.8.0,<2
-    - tqdm >=4.15,<5
-    - XlsxWriter >=1.2.8,<4
-    - rdt >=0.6.1,<0.7
-    - sdmetrics >=0.4.1,<0.5
-    - sdv >=0.13.0
-  run:
-    - python >=3.6,<3.10
-    - appdirs >=1.3,<2
-    - boto3 >=1.15.0,<2
-    - botocore >=1.18,<2
-    - humanfriendly >=8.2,<11
-    - numpy >=1.18.0,<2
-    - pandas >=1.1.3,<2
-    - pomegranate >=0.13.4,<15
-    - psutil >=5.7,<6
-    - scikit-learn >=0.24,<2
-    - tabulate >=0.8.3,<0.9
-    - pytorch >=1.8.0,<2
-    - tqdm >=4.15,<5
-    - XlsxWriter >=1.2.8,<4
-    - rdt >=0.6.1,<0.7
-    - sdmetrics >=0.4.1,<0.5
-    - sdv >=0.13.0
-
-about:
-  home: "https://github.com/sdv-dev/SDGym"
-  license: BUSL-1.1
-  license_family:
-  license_file: LICENSE
-  summary: "A framework to benchmark the performance of synthetic data generators for non-temporal tabular data"
-  doc_url:
-  dev_url:
-
-extra:
-  recipe-maintainers:
-    - sdv-dev
