+++ sdgym/__init__.py
@@ -14,9 +14,9 @@ import logging
 
 from sdgym import benchmark, synthesizers
 from sdgym.benchmark import benchmark_single_table
-from sdgym.collect import collect_results
+from sdgym.cli.collect import collect_results
+from sdgym.cli.summary import make_summary_spreadsheet
 from sdgym.datasets import load_dataset
-from sdgym.summary import make_summary_spreadsheet
 
 # Clear the logging wrongfully configured by tensorflow/absl
 list(map(logging.root.removeHandler, logging.root.handlers))
